requirements.txt
absl-py==0.12.0
aiofiles==0.6.0
attr==0.3.1
attrs==21.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coremltools==4.1
cycler==0.10.0
Cython==0.29.23
dyNET==2.1.2
fastapi==0.65.1
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
h11==0.12.0
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mpmath==1.2.1
numpy==1.20.3
oauthlib==3.1.0
onnx==1.9.0
onnxruntime-gpu==1.7.0
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydantic==1.8.2
pydicom==2.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
sklearn==0.0
starlette==0.14.2
sympy==1.8
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
Werkzeug==2.0.0
